integrate API with React app
This commit replaces the old React app API with the
new Node API, using appropriate routes to fetch data
from the TMDB API and passing them to the React app.

It implements the following new routes, which interact
with the React app (i.e. ALL old API calls are replaced):
- /api/movies/tmdb/genres
- /api/movies/tmdb/discover
- /api/movies/tmdb/trending/week
- /api/movies/tmdb/trending/today
- /api/movies/tmdb/:id
- /api/movies/tmdb/:id/reviews
- /api/movies/tmdb/:id/images
- /api/movies/tmdb/:id/credits
- /api/movies/tmdb/person/:id

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>