add movie details schema and mongodb functionality
This commit adds the movie details schema (slightly
different from the movie schema, used for storing
more specific details abput a movie).

Movie details works with the DB in the same way the
person details does i.e. adds the movie to the DB
on API call, if the movie is not already added.

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>