refactor favourites feature
This commit integrates the favourites functionality
from the previous react app into the new node app with
MongoDB use.

The app now stores movies as favourites within the user's
document of the DB (as an array). This array is then altered
as required (add/remove movies) and the display changes
dynamically with this functionality as before.

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>