implement change password feature
This commit implements a way for users to change their
passwords after logging into their account.

The hashing function was also refactored to enable this
functionality.

This is done through a new page for managing the account,
where the user now also logs out rather than from the
site header.

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>