configure CI/CD as required
This commit refactores the .gitlab-ci.yml file in
order to add a job deploy_production during the
deploy stage, which automatically deploys the
app to the production server when the main branch
is pushed i.e. implementing Continuous deployment.

The app may be deployed to the production server
on manual trigger of the job on the develop branch
i.e. implementing Continuous Delivery.

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>